isolate handling session  methods
